restructured theme application to use useTheme hook, renamed theme slice to preferences
