+++ app/src/index.css
@@ -28,7 +28,7 @@
   --hover: white;
   --text-color: #bbb;
   --box-size: 20px;
-  --box-background: #666;
+  --box-background: #444;
   --border: 1px solid #999;
   --radius: 4px;
   --icon-hover-background: #555;
